requirements.txt
#--- Core Application & Web Framework ---
fastapi==0.116.1
uvicorn[standard]==0.33.0
pydantic[email]==2.10.6
pydantic==2.10.6
pydantic-settings==2.4.0
loguru==0.7.2
python-multipart==0.0.9

#--- Database & ORM ---
sqlalchemy==2.0.33
psycopg2-binary==2.9.10
alembic==1.13.2

#--- Security & Authentication ---
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

#--- Asynchronous Tasks & Caching ---
celery==5.4.0
pika>=1.0.0
redis==5.0.7

#--- Testing ---
pytest==8.3.3
pytest-asyncio==0.23.8
anyio==4.3.0
httpx==0.27.0
locust==2.32.3

# Harmless package to force cache invalidation
pytz==2024.1
